feat: update accept ride request models and improve navigation handling

+++ lib/domain/fcm/models/accept_ride_request/accept_ride_request_vehicle.dart
@@ -8,7 +8,7 @@ class AcceptRideRequestVehicle {
   String? vehicleId;
   String? name;
   @JsonKey(name: 'fuel_consumed')
-  int? fuelConsumed;
+  double? fuelConsumed;
   @JsonKey(name: 'license_plate')
   String? licensePlate;
